requirements.txt
google-auth==2.0.2
gspread==4.0.1
gspread-dataframe==3.2.0
pandas==1.3.3
numpy==1.21.2
requests==2.26.0